requirements.txt
--index-url https://pypi.python.org/simple
--extra-index-url https://pypi-reader:ThOgrAVultiA@artifacts.ua-ecm.com/artifactory/api/pypi/ua-pypi/simple
-e git://github.com/mozilla/elasticutils.git
pony
arrow==0.7.0
boto3==1.2.6
jira
botocore==1.4.49
Django==1.9.2
Mako==1.0.3
oauthlib==1.0.3
pathspec==0.3.4
pyfluence==0.1.1
PyGithub==1.36
pymongo==3.2.1
python-dateutil==2.5.3
pytz==2015.7
urllib3==1.21.1
psycopg2==2.6.1
PyYAML==3.12
munch
